+++ src/main/java/msifeed/misca/combat/rules/Dices.java
@@ -5,12 +5,12 @@ import java.util.Random;
 public class Dices {
     private static final Random rand = new Random();
 
-    public static boolean check(float chance) {
+    public static boolean check(double chance) {
         return rand.nextFloat() < chance;
     }
 
-    public static float checkFloat(float chance) {
-        return rand.nextFloat() < chance ? 1 : 0;
+    public static double checkFloat(double chance) {
+        return rand.nextDouble() < chance ? 1 : 0;
     }
 
     public static int d20() {
